feat(app): add tkinter dependency and create_initial_grid function
The tkinter dependency was added to the Pipfile and the create_initial_grid function was added to the main.py file. This function creates a graphical interface to allow the user to create the initial grid for the Game of Life simulation. The user can click on the cells to toggle them between alive and dead. Once the user is satisfied with the initial grid, they can click the "Start" button to begin the simulation.